fix(app): add missing `EntryProvider`s to `KiteApp`
