+++ app/src/main/kotlin/com/potatosheep/kite/app/ui/KiteAppState.kt
@@ -6,14 +6,13 @@ import com.potatosheep.kite.app.navigation.TopLevelDestination
 import com.potatosheep.kite.app.navigation.TopLevelNavKey
 import com.potatosheep.kite.core.navigation.NavigationState
 import com.potatosheep.kite.core.navigation.rememberNavigationState
-import com.potatosheep.kite.feature.feed.api.navigation.FeedNavKey
 import com.potatosheep.kite.feature.onboarding.api.navigation.OnboardingNavKey
 
 @Composable
 fun rememberAppState(
     shouldShowOnboarding: Boolean,
 ): KiteAppState {
-    val startRoute = if (shouldShowOnboarding) OnboardingNavKey else FeedNavKey
+    val startRoute = if (shouldShowOnboarding) OnboardingNavKey else TopLevelNavKey
     val navigationState = rememberNavigationState(startRoute, setOf(TopLevelNavKey));
 
     return remember {
